+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleConsentTests.java
@@ -3,19 +3,26 @@ package ly.count.android.sdk;
 import android.app.Application;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.List;
+import org.json.JSONException;
+import org.json.JSONObject;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Mockito;
 
 import static androidx.test.InstrumentationRegistry.getContext;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyZeroInteractions;
 
 @RunWith(AndroidJUnit4.class)
 public class ModuleConsentTests {
@@ -45,9 +52,7 @@ public class ModuleConsentTests {
     public void tearDown() {
     }
 
-
-
-    Countly helperCreateAndInitCountly(){
+    Countly helperCreateAndInitCountly() {
         return null;
     }
 
@@ -175,42 +180,121 @@ public class ModuleConsentTests {
         Assert.assertEquals("clicks", Countly.CountlyFeatureNames.clicks);
     }
 
+    /**
+     * Require no consent at init time and provide no consent values.
+     * Does not provide any location or anything else
+     * No requests should be created.
+     * There should be no interactions with the mock
+     */
     @Test
-    public void initTimeSetConsentRQ() {
+    public void initTimeNoConsentRequiredRQ() {
         RequestQueueProvider rqp = mock(RequestQueueProvider.class);
-        Countly mCountly = new Countly().init(TestUtils.createConsentCountlyConfig(true, null, null, rqp));
-
-        //this should send no consent changes and empty location
-        verify(rqp, times(0)).sendConsentChanges(any(String.class));
-        verify(rqp, times(1)).sendLocation(true, null, null, null, null);
+        Countly mCountly = new Countly().init(TestUtils.createConsentCountlyConfig(false, null, null, rqp));
+        verifyZeroInteractions(rqp);
     }
 
-    @Test
-    public void initTimeSetConsentRQ_2() {
-        RequestQueueProvider rqp = mock(RequestQueueProvider.class);
-        Countly mCountly = new Countly().init(TestUtils.createConsentCountlyConfig(true, new String[] { Countly.CountlyFeatureNames.clicks }, null, rqp));
+    void verifyLocationValuesInRQMock(int count, Boolean enabled, String countryCode, String city, String location, String ip, RequestQueueProvider rqp) {
+        ArgumentCaptor<Boolean> acLocationDisabled = ArgumentCaptor.forClass(Boolean.class);
+        ArgumentCaptor<String> acCountryCode = ArgumentCaptor.forClass(String.class);
+        ArgumentCaptor<String> acCity = ArgumentCaptor.forClass(String.class);
+        ArgumentCaptor<String> acGps = ArgumentCaptor.forClass(String.class);
+        ArgumentCaptor<String> acIp = ArgumentCaptor.forClass(String.class);
+        verify(rqp, times(count)).sendLocation(acLocationDisabled.capture(), acCountryCode.capture(), acCity.capture(), acGps.capture(), acIp.capture());
 
-        //this should send consent changes and empty location
-        verify(rqp, times(1)).sendConsentChanges("{\"clicks\":true}");
-        verify(rqp, times(1)).sendLocation(true, null, null, null, null);
+        if (count == 0) {
+            return;
+        }
+
+        Assert.assertEquals(enabled, acLocationDisabled.getValue());
+        Assert.assertEquals(countryCode, acCountryCode.getValue());
+        Assert.assertEquals(city, acCity.getValue());
+        Assert.assertEquals(location, acGps.getValue());
+        Assert.assertEquals(ip, acIp.getValue());
+    }
+
+    void verifyConsentValuesInRQMock(int count, String[] valuesTrue, String[] valuesFalse, RequestQueueProvider rqp) throws JSONException {
+        ArgumentCaptor<String> consentChanges = ArgumentCaptor.forClass(String.class);
+        verify(rqp, times(count)).sendConsentChanges(consentChanges.capture());
+
+        String changes = consentChanges.getValue();
+        Assert.assertNotNull(changes);
+
+        JSONObject jObj = new JSONObject(changes);
+
+        Assert.assertEquals(usedFeatureNames.length, jObj.length());
+        Assert.assertEquals(usedFeatureNames.length, valuesTrue.length + valuesFalse.length);
+
+        for (String v : valuesTrue) {
+            Assert.assertTrue((Boolean) jObj.get(v));
+        }
+
+        for (String v : valuesFalse) {
+            Assert.assertFalse((Boolean) jObj.get(v));
+        }
+    }
+
+    String[] subtractConsentFromArray(String[] input, String[] subtraction) {
+        ArrayList<String> res = new ArrayList<>();
+
+        for(String v:input) {
+            boolean contains = false;
+            for(String sv:subtraction) {
+                if(sv.equals(v)) {
+                    contains = true;
+                    break;
+                }
+            }
+
+            if(!contains) {
+                res.add(v);
+            }
+        }
+
+        return (String[]) res.toArray();
     }
 
     /**
-     * Setting single consent value during init.
-     * That produces appropriate consent request and since it's the location request, does not clear location
+     * Consent required at init time but no consent values are given
+     *
+     * This should create a request of false consent values and a location request with removed location
      */
     @Test
-    public void initTimeSetConsentRQ_3() {
+    public void initTimeNoConsentGivenRQ() throws JSONException {
         RequestQueueProvider rqp = mock(RequestQueueProvider.class);
-        Countly mCountly = new Countly().init(TestUtils.createConsentCountlyConfig(true, new String[] { Countly.CountlyFeatureNames.location }, null, rqp));
+        Countly mCountly = new Countly().init(TestUtils.createConsentCountlyConfig(true, null, null, rqp));
+        Assert.assertEquals(2, Mockito.mockingDetails(rqp).getInvocations().size());
 
-        //this should send consent changes and empty location
-        verify(rqp, times(1)).sendConsentChanges("{\"location\":true}");
-        verify(rqp, times(0)).sendLocation(true, null, null, null, null);
+        verifyLocationValuesInRQMock(1, true, null, null, null, null, rqp);
+        verifyConsentValuesInRQMock(1, new String[] {}, usedFeatureNames, rqp);
     }
+
+    //@Test
+    //public void initTimeSetConsentRQ_2() {
+    //    RequestQueueProvider rqp = mock(RequestQueueProvider.class);
+    //    Countly mCountly = new Countly().init(TestUtils.createConsentCountlyConfig(true, new String[] { Countly.CountlyFeatureNames.clicks }, null, rqp));
+    //
+    //    //this should send consent changes and empty location
+    //    verify(rqp, times(1)).sendConsentChanges("{\"clicks\":true}");
+    //    verify(rqp, times(1)).sendLocation(true, null, null, null, null);
+    //}
+    //
+    ///**
+    // * Setting single consent value during init.
+    // * That produces appropriate consent request and since it's the location request, does not clear location
+    // */
+    //@Test
+    //public void initTimeSetConsentRQ_3() {
+    //    RequestQueueProvider rqp = mock(RequestQueueProvider.class);
+    //    Countly mCountly = new Countly().init(TestUtils.createConsentCountlyConfig(true, new String[] { Countly.CountlyFeatureNames.location }, null, rqp));
+    //
+    //    //this should send consent changes and empty location
+    //    verify(rqp, times(1)).sendConsentChanges("{\"location\":true}");
+    //    verify(rqp, times(0)).sendLocation(true, null, null, null, null);
+    //}
 
     // TODO test that makes sure that the consent change request is created correctly
     // TODO test that makes sure that the consent change request is not created for duplicate triggers
     // TODO make sure that the consent request is correctly created after init
     // TODO test that consent given / removed triggers are played correctly
+    // TODO ?make tests for formatConsentState
 }