+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   * Added new calls to report attribution.
   * When recording internal events with 'recordEvent', the respective feature consent will now be checked instead of the 'events' consent. 
   * Adding an init time configuration option to set the maximum request queue size
+  * Consent changes will now send the whole consent state and not just the "delta"
+
   * Updating the "compileSdkVersion" and "targetSdkVersion" to "31".
   * Updating the "buildToolsVersion" to "31.0.0".